nutrition animation/movement complete

+++ app/src/main/java/com/example/food4u/Ball.java
@@ -5,23 +5,23 @@ public class Ball {
     public static final double MAX_SPEED = 50;
     public static final int WIDTH = 10;
     public static final int HEIGHT = 10;
-    private double x;
-    private double y;
+    private float x;
+    private float y;
     private double speed;
     private double angle;
 
-    public Ball(double x, double y) {
+    public Ball(float x, float y) {
         this.x = x;
         this.y = y;
         angle = 180;
         speed = 4;
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
@@ -33,11 +33,11 @@ public class Ball {
         return speed;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
